Separate methods for handling local attachments into separate class
This leaves the 'Attachment' class free for a pure representation of an attachment
